fix: move RealtimeAdapter contract to @tanstack/ai so provider packages don't depend on ai-client
Provider adapter packages (ai-openai, ai-elevenlabs, ai-grok) are usable
server-side and must not pull in the client-only @tanstack/ai-client. The only
thing their realtime adapters needed from it were the RealtimeAdapter /
RealtimeConnection type shapes.

Those interfaces now live in @tanstack/ai — the shared layer both providers and
@tanstack/ai-client already depend on, alongside every other realtime type — and
are exported from @tanstack/ai and @tanstack/ai/client. @tanstack/ai-client
re-exports them unchanged, so `import { RealtimeAdapter } from '@tanstack/ai-client'`
keeps working. The previously-duplicated local contract + drift test in ai-grok
are removed in favor of the single shared definition.

@tanstack/ai-client is no longer a dependency (peer or otherwise) of any provider
package.
