+++ packages/ai-elevenlabs/package.json
@@ -63,12 +63,10 @@
     "@tanstack/ai-utils": "workspace:*"
   },
   "peerDependencies": {
-    "@tanstack/ai": "workspace:^",
-    "@tanstack/ai-client": "workspace:^"
+    "@tanstack/ai": "workspace:^"
   },
   "devDependencies": {
     "@tanstack/ai": "workspace:*",
-    "@tanstack/ai-client": "workspace:*",
     "@vitest/coverage-v8": "4.0.14"
   }
 }
