+++ packages/ai-grok/package.json
@@ -65,7 +65,6 @@
   },
   "devDependencies": {
     "@tanstack/ai": "workspace:*",
-    "@tanstack/ai-client": "workspace:*",
     "@vitest/coverage-v8": "4.0.14",
     "vite": "^7.3.3"
   },
