+++ packages/ai-openai/src/realtime/adapter.ts
@@ -3,6 +3,8 @@ import { buildSessionUpdate } from './session-update'
 import type {
   AnyClientTool,
   AudioVisualization,
+  RealtimeAdapter,
+  RealtimeConnection,
   RealtimeEvent,
   RealtimeEventHandler,
   RealtimeMessage,
@@ -11,7 +13,6 @@ import type {
   RealtimeToken,
 } from '@tanstack/ai'
 import type { InternalLogger } from '@tanstack/ai/adapter-internals'
-import type { RealtimeAdapter, RealtimeConnection } from '@tanstack/ai-client'
 import type { OpenAIRealtimeOptions } from './types'
 
 const OPENAI_REALTIME_URL = 'https://api.openai.com/v1/realtime'
